+++ HISTORY.md
@@ -1,12 +1,18 @@
 # History
 
+## UNRELEASED
+
+- Added PLAUSIBLE_BASE_URL settings option to make it possible to use the project with self-hosted Plausible installations. The default value is https://plausible.io (use the cloud version.)
+- Made PLAUSIBLE_DOMAIN settings optional. If the value is not set, the domain name is taken from the request.
+- Added more tests.
+
 ## 0.1.2 (2022-04-25)
 
-- Fix app config
+- Fixed app config
 
 ## 0.1.1 (2022-04-25)
 
-- Fix project metadata
+- Fixed project metadata
 
 ## 0.1.0 (2022-04-25)